+++ php7/ext/oci8/tests/bind_misccoltypes_errs.phpt
@@ -1,22 +1,18 @@
 --TEST--
 Bind miscellaneous column types and generating errors
---EXTENSIONS--
-oci8
 --SKIPIF--
-<?php
-require_once 'skipifconnectfailure.inc';
-?>
+<?php if (!extension_loaded('oci8')) die ("skip no oci8 extension"); ?>
 --FILE--
 <?php
 
-require __DIR__.'/connect.inc';
+require(__DIR__.'/connect.inc');
 
 // Initialization
 
 $stmtarray = array(
-    "drop table bind_misccoltypes_errs_tab",
+	"drop table bind_misccoltypes_errs_tab",
 
-    "create table bind_misccoltypes_errs_tab (
+	"create table bind_misccoltypes_errs_tab (
         id                number,
         char_t            char(1),
         char_t10          char(10),
@@ -112,7 +108,7 @@ check_col($c, 'varchar2_t10', 7);
 // Clean up
 
 $stmtarray = array(
-    "drop table bind_misccoltypes_errs_tab"
+	"drop table bind_misccoltypes_errs_tab"
 );
 
 oci8_test_sql_execute($c, $stmtarray);
@@ -120,6 +116,8 @@ oci8_test_sql_execute($c, $stmtarray);
 oci_close($c);
 
 ?>
+===DONE===
+<?php exit(0); ?>
 --EXPECTF--
 Test 1 insert numbers
 
@@ -168,3 +166,4 @@ array(1) {
   array(0) {
   }
 }
+===DONE===
