+++ php7/ext/oci8/tests/bind_sqltint.phpt
@@ -1,20 +1,16 @@
 --TEST--
 Bind with SQLT_INT
---EXTENSIONS--
-oci8
 --SKIPIF--
-<?php
-require_once 'skipifconnectfailure.inc';
-?>
+<?php if (!extension_loaded('oci8')) die ("skip no oci8 extension"); ?>
 --FILE--
 <?php
 
-require __DIR__.'/connect.inc';
+require(__DIR__.'/connect.inc');
 
 // Initialization
 
 $stmtarray = array(
-    "drop table bind_sqltint_tab",
+	"drop table bind_sqltint_tab",
 
     "create table bind_sqltint_tab (
         id                number,
@@ -212,12 +208,14 @@ var_dump($p2);
 // Clean up
 
 $stmtarray = array(
-    "drop table bind_sqltint_tab"
+	"drop table bind_sqltint_tab"
 );
 
 oci8_test_sql_execute($c, $stmtarray);
 
 ?>
+===DONE===
+<?php exit(0); ?>
 --EXPECT--
 TEST141 wrong bind type SQLT_INT
 array(1) {
@@ -335,3 +333,4 @@ int(4444)
 
 TEST159
 int(5555)
+===DONE===
