+++ php7/ext/oci8/tests/error_bind.phpt
@@ -1,15 +1,11 @@
 --TEST--
 Test some oci_bind_by_name error conditions
---EXTENSIONS--
-oci8
 --SKIPIF--
-<?php
-require_once 'skipifconnectfailure.inc';
-?>
+<?php if (!extension_loaded('oci8')) die("skip no oci8 extension"); ?>
 --FILE--
 <?php
 
-require __DIR__.'/connect.inc';
+require(__DIR__.'/connect.inc');
 
 $drop = "drop table bind_test";
 $statement = oci_parse($c, $drop);
