+++ php7/ext/oci8/tests/xmltype_02.phpt
@@ -1,24 +1,21 @@
 --TEST--
 Basic XMLType test #2
---EXTENSIONS--
-simplexml
-oci8
 --SKIPIF--
 <?php
-require_once 'skipifconnectfailure.inc';
 $target_dbs = array('oracledb' => true, 'timesten' => false);  // test runs on these DBs
-require __DIR__.'/skipif.inc';
+require(__DIR__.'/skipif.inc');
+if (!extension_loaded("simplexml")) die ("skip no simplexml extension");
 ?>
 --FILE--
 <?php
 
-require __DIR__.'/connect.inc';
+require(__DIR__.'/connect.inc');
 
 // Initialization
 
 $stmtarray = array(
-    "drop table xmltype_02_tab",
-    "create table xmltype_02_tab (warehouse_id number, warehouse_spec xmltype)",
+	"drop table xmltype_02_tab",
+	"create table xmltype_02_tab (warehouse_id number, warehouse_spec xmltype)",
 );
 
 oci8_test_sql_execute($c, $stmtarray);
@@ -104,17 +101,19 @@ $row[0]->free();
 // Clean up
 
 $stmtarray = array(
-    "drop table xmltype_02_tab"
+	"drop table xmltype_02_tab"
 );
 
 oci8_test_sql_execute($c, $stmtarray);
 
 ?>
+===DONE===
+<?php exit(0); ?>
 --EXPECTF--
 Test 1 Insert new XML data using a temporary CLOB
 array(1) {
   [0]=>
-  object(OCILob)#%d (1) {
+  object(OCI-Lob)#%d (1) {
     ["descriptor"]=>
     resource(%d) of type (oci8 descriptor)
   }
@@ -179,3 +178,4 @@ string(%d) "<?xml version="1.0"?>
 %sVClearance>10</VClearance>
 </Warehouse>
 "
+===DONE===
