+++ php7/ext/zip/tests/oo_setcompression.phpt
@@ -1,19 +1,21 @@
 --TEST--
 setCompressionName and setCompressionIndex methods
---EXTENSIONS--
-zip
+--SKIPIF--
+<?php
+if (!extension_loaded('zip')) die('skip');
+?>
 --FILE--
 <?php
 $tmpfile = __DIR__ . '/oo_setcompression.zip';
 
 if (file_exists($tmpfile)) {
-    unlink($tmpfile);
+	unlink($tmpfile);
 }
 
 // generate the ZIP file
 $zip = new ZipArchive;
 if ($zip->open($tmpfile, ZipArchive::CREATE) !== TRUE) {
-    exit('failed');
+	exit('failed');
 }
 $txt = file_get_contents(__FILE__);
 $zip->addFromString('entry1.txt', $txt);
@@ -33,18 +35,18 @@ var_dump($zip->setCompressionIndex(5, ZipArchive::CM_DEFLATE));
 var_dump($zip->setCompressionIndex(6, ZipArchive::CM_DEFAULT));
 
 if (!$zip->close()) {
-    exit('failed');
+	exit('failed');
 }
 
 
 // check the ZIP file
 $zip = zip_open($tmpfile);
 if (!is_resource($zip)) {
-    exit('failed');
+	exit('failed');
 }
 
 while ($e = zip_read($zip)) {
-    echo zip_entry_name($e) . ': ' . zip_entry_compressionmethod($e) . "\n";
+	echo zip_entry_name($e) . ': ' . zip_entry_compressionmethod($e) . "\n";
 }
 zip_close($zip);
 ?>
@@ -53,65 +55,17 @@ zip_close($zip);
 $tmpfile = __DIR__ . '/oo_setcompression.zip';
 unlink($tmpfile);
 ?>
---EXPECTF--
+--EXPECT--
 bool(true)
 bool(true)
 bool(true)
 bool(true)
 bool(true)
 bool(true)
-
-Deprecated: Function zip_open() is deprecated in %s on line %d
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry1.txt: deflated
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry2.txt: deflated
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 dir/entry3.txt: stored
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry4.txt: deflated
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry5.txt: stored
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry6.txt: deflated
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_name() is deprecated in %s on line %d
-
-Deprecated: Function zip_entry_compressionmethod() is deprecated in %s on line %d
 entry7.txt: deflated
-
-Deprecated: Function zip_read() is deprecated in %s on line %d
-
-Deprecated: Function zip_close() is deprecated in %s on line %d
