+++ php7/ext/zlib/tests/bug61139.phpt
@@ -1,15 +1,18 @@
 --TEST--
 Bug #61139 (gzopen leaks when specifying invalid mode)
---EXTENSIONS--
-zlib
+--SKIPIF--
+<?php
+if (!extension_loaded('zlib')) {
+	die('skip - zlib extension not loaded');
+}
+?>
 --FILE--
 <?php
 
 gzopen('someFile', 'c');
-?>
 --CLEAN--
 <?php
-    unlink('someFile');
+	unlink('someFile');
 ?>
 --EXPECTF--
 Warning: gzopen(): gzopen failed in %s on line %d
